UID Grab & Roles Changes
UID Grabs now output 2 separate files, one with just uids, the other with JSON objects with the uid and name of player.  That way there is no way of the two files needing to rely on each other.

Players with roles in their names' messages will now be ignored by racers.