IP Logger Changed
IP logger now checks if it's the first time logging if another rider has a matched IP. This is for the scenario someone is using a smurf key. Also added names to logging instead of just uids. Names are checked each time and updated if name doesn't match last name they were logged.